aamp/requirements.txt
bootstrap-admin==0.3.6
diff-match-patch==20121119
Django==1.9.2
django-allauth==0.24.1
django-appconf==1.0.1
django-categories==1.4
django-ckeditor==5.0.3
django-classy-tags==0.7.2
django-cms==3.2.2
django-easy-pdf==0.1.0
django-filer==1.1.1
django-formtools==1.0
django-geoposition==0.2.2
django-import-export==0.4.2
django-meta==0.3.2
django-meta-mixin==0.2.1
django-mptt==0.8.3
django-polymorphic==0.8.1
django-robots==2.0
django-sekizai==0.9.0
django-treebeard==4.0
django-wysiwyg==0.7.1
djangocms-admin-style==1.1.0
DjangoSEO==1.0
easy-thumbnails==2.3
grappelli-side-menu==0.1.3
html5lib==0.9999999
httplib2==0.9.2
oauth2==1.9.0.post1
oauthlib==1.0.3
pdfcrowd==2.7
pdfkit==0.5.0
Pillow==3.1.1
psycopg2==2.6.1
PyPDF2==1.25.1
python-openid==2.2.5
reportlab==3.3.0
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
South==1.0.2
tablib==0.11.1
unicode-slugify==0.1.1
Unidecode==0.4.19
wheel==0.29.0
xhtml2pdf==0.0.6
